gather app utils in `/src`
