+++ app/src/utils/snarkjs.ts
